requirements.txt
flask==3.0.0
requests
GitPython
discord
python-dotenv